requirements.txt
discord.py>=2.1.1
request>=2.28.2